+++ encoding/codecv3_types.go
@@ -12,205 +12,6 @@ import (
 	"github.com/scroll-tech/go-ethereum/crypto/kzg4844"
 )
 
-// daBatchV0 contains metadata about a batch of DAChunks.
-type daBatchV0 struct {
-	version                uint8
-	batchIndex             uint64
-	l1MessagePopped        uint64
-	totalL1MessagePopped   uint64
-	dataHash               common.Hash
-	parentBatchHash        common.Hash
-	skippedL1MessageBitmap []byte
-}
-
-// newDABatchV0 is a constructor for daBatchV0.
-func newDABatchV0(version uint8, batchIndex, l1MessagePopped, totalL1MessagePopped uint64, dataHash, parentBatchHash common.Hash, skippedL1MessageBitmap []byte) *daBatchV0 {
-	return &daBatchV0{
-		version:                version,
-		batchIndex:             batchIndex,
-		l1MessagePopped:        l1MessagePopped,
-		totalL1MessagePopped:   totalL1MessagePopped,
-		dataHash:               dataHash,
-		parentBatchHash:        parentBatchHash,
-		skippedL1MessageBitmap: skippedL1MessageBitmap,
-	}
-}
-
-// Encode serializes the DABatch into bytes.
-func (b *daBatchV0) Encode() []byte {
-	batchBytes := make([]byte, 89+len(b.skippedL1MessageBitmap))
-	batchBytes[0] = b.version
-	binary.BigEndian.PutUint64(batchBytes[1:], b.batchIndex)
-	binary.BigEndian.PutUint64(batchBytes[9:], b.l1MessagePopped)
-	binary.BigEndian.PutUint64(batchBytes[17:], b.totalL1MessagePopped)
-	copy(batchBytes[25:], b.dataHash[:])
-	copy(batchBytes[57:], b.parentBatchHash[:])
-	copy(batchBytes[89:], b.skippedL1MessageBitmap[:])
-	return batchBytes
-}
-
-// Hash computes the hash of the serialized DABatch.
-func (b *daBatchV0) Hash() common.Hash {
-	bytes := b.Encode()
-	return crypto.Keccak256Hash(bytes)
-}
-
-// Blob returns the blob of the batch.
-func (b *daBatchV0) Blob() *kzg4844.Blob {
-	return nil
-}
-
-// BlobVersionedHashes returns the blob versioned hashes of the batch.
-func (b *daBatchV0) BlobVersionedHashes() []common.Hash {
-	return nil
-}
-
-// BlobBytes returns the blob bytes of the batch.
-func (b *daBatchV0) BlobBytes() []byte {
-	return nil
-}
-
-// BlobDataProofForPointEvaluation computes the abi-encoded blob verification data.
-func (b *daBatchV0) BlobDataProofForPointEvaluation() ([]byte, error) {
-	return nil, nil
-}
-
-// Version returns the version of the DABatch.
-func (b *daBatchV0) Version() uint8 {
-	return b.version
-}
-
-// SkippedL1MessageBitmap returns the skipped L1 message bitmap of the DABatch.
-func (b *daBatchV0) SkippedL1MessageBitmap() []byte {
-	return b.skippedL1MessageBitmap
-}
-
-// DataHash returns the data hash of the DABatch.
-func (b *daBatchV0) DataHash() common.Hash {
-	return b.dataHash
-}
-
-// daBatchV1 contains metadata about a batch of DAChunks.
-type daBatchV1 struct {
-	daBatchV0
-
-	blobVersionedHash common.Hash
-	blob              *kzg4844.Blob
-	z                 *kzg4844.Point
-}
-
-// newDABatchV1 is a constructor for daBatchV1.
-func newDABatchV1(version uint8, batchIndex, l1MessagePopped, totalL1MessagePopped uint64, dataHash, parentBatchHash, blobVersionedHash common.Hash, skippedL1MessageBitmap []byte, blob *kzg4844.Blob, z *kzg4844.Point) *daBatchV1 {
-	return &daBatchV1{
-		daBatchV0: daBatchV0{
-			version:                version,
-			batchIndex:             batchIndex,
-			l1MessagePopped:        l1MessagePopped,
-			totalL1MessagePopped:   totalL1MessagePopped,
-			dataHash:               dataHash,
-			parentBatchHash:        parentBatchHash,
-			skippedL1MessageBitmap: skippedL1MessageBitmap,
-		},
-		blobVersionedHash: blobVersionedHash,
-		blob:              blob,
-		z:                 z,
-	}
-}
-
-// Encode serializes the DABatch into bytes.
-func (b *daBatchV1) Encode() []byte {
-	batchBytes := make([]byte, 121+len(b.skippedL1MessageBitmap))
-	batchBytes[0] = b.version
-	binary.BigEndian.PutUint64(batchBytes[1:], b.batchIndex)
-	binary.BigEndian.PutUint64(batchBytes[9:], b.l1MessagePopped)
-	binary.BigEndian.PutUint64(batchBytes[17:], b.totalL1MessagePopped)
-	copy(batchBytes[25:], b.dataHash[:])
-	copy(batchBytes[57:], b.blobVersionedHash[:])
-	copy(batchBytes[89:], b.parentBatchHash[:])
-	copy(batchBytes[121:], b.skippedL1MessageBitmap[:])
-	return batchBytes
-}
-
-// Hash computes the hash of the serialized DABatch.
-func (b *daBatchV1) Hash() common.Hash {
-	bytes := b.Encode()
-	return crypto.Keccak256Hash(bytes)
-}
-
-// BlobDataProof computes the abi-encoded blob verification data.
-func (b *daBatchV1) BlobDataProof() ([]byte, error) {
-	if b.blob == nil {
-		return nil, errors.New("called BlobDataProof with empty blob")
-	}
-	if b.z == nil {
-		return nil, errors.New("called BlobDataProof with empty z")
-	}
-
-	commitment, err := kzg4844.BlobToCommitment(b.blob)
-	if err != nil {
-		return nil, errors.New("failed to create blob commitment")
-	}
-
-	proof, y, err := kzg4844.ComputeProof(b.blob, *b.z)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create KZG proof at point, err: %w, z: %v", err, hex.EncodeToString(b.z[:]))
-	}
-
-	return BlobDataProofFromValues(*b.z, y, commitment, proof), nil
-}
-
-// Blob returns the blob of the batch.
-func (b *daBatchV1) Blob() *kzg4844.Blob {
-	return b.blob
-}
-
-// BlobVersionedHashes returns the blob versioned hashes of the batch.
-func (b *daBatchV1) BlobVersionedHashes() []common.Hash {
-	return []common.Hash{b.blobVersionedHash}
-}
-
-// BlobBytes returns the blob bytes of the batch.
-func (b *daBatchV1) BlobBytes() []byte {
-	return nil
-}
-
-// BlobDataProofForPointEvaluation computes the abi-encoded blob verification data.
-func (b *daBatchV1) BlobDataProofForPointEvaluation() ([]byte, error) {
-	if b.blob == nil {
-		return nil, errors.New("called BlobDataProofForPointEvaluation with empty blob")
-	}
-	if b.z == nil {
-		return nil, errors.New("called BlobDataProofForPointEvaluation with empty z")
-	}
-
-	commitment, err := kzg4844.BlobToCommitment(b.blob)
-	if err != nil {
-		return nil, errors.New("failed to create blob commitment")
-	}
-
-	proof, y, err := kzg4844.ComputeProof(b.blob, *b.z)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create KZG proof at point, err: %w, z: %v", err, hex.EncodeToString(b.z[:]))
-	}
-
-	return BlobDataProofFromValues(*b.z, y, commitment, proof), nil
-}
-
-// Version returns the version of the DABatch.
-func (b *daBatchV1) Version() uint8 {
-	return b.version
-}
-
-// SkippedL1MessageBitmap returns the skipped L1 message bitmap of the DABatch.
-func (b *daBatchV1) SkippedL1MessageBitmap() []byte {
-	return b.skippedL1MessageBitmap
-}
-
-// DataHash returns the data hash of the DABatch.
-func (b *daBatchV1) DataHash() common.Hash {
-	return b.dataHash
-}
-
 // daBatchV2 contains metadata about a batch of DAChunks.
 type daBatchV2 struct {
 	daBatchV0
@@ -356,11 +157,6 @@ func (b *daBatchV2) BlobDataProofForPointEvaluation() ([]byte, error) {
 // Blob returns the blob of the batch.
 func (b *daBatchV2) Blob() *kzg4844.Blob {
 	return b.blob
-}
-
-// BlobVersionedHashes returns the blob versioned hashes of the batch.
-func (b *daBatchV2) BlobVersionedHashes() []common.Hash {
-	return []common.Hash{b.blobVersionedHash}
 }
 
 // BlobBytes returns the blob bytes of the batch.
